Make getComputerChoice, getPlayerChoice and playRound functions
Created three functions, the first, getComputerChoice, makes the
machine pick a random number from cero to 2 and pick it's corresponding
word from an array containing "rock", "paper", and "scissors" strings.

The second function, getPLayerChoice, uses prompt() to ask the user
for input, the input is assigned to a new variable and it is later returned
in all lowercase.

The third function, playRound assigns the return values of the first funcitons to two new variables
corresponding to the player and computer selection. It is still incomplete.